+++ src/main/java/com/runninghi/bookmarkfolder/query/application/service/QueryBookmarkFolderService.java
@@ -3,17 +3,14 @@ package com.runninghi.bookmarkfolder.query.application.service;
 import com.runninghi.bookmarkfolder.command.domain.aggregate.entity.BookmarkFolder;
 import com.runninghi.bookmarkfolder.query.application.dto.request.FindFolderRequest;
 import com.runninghi.bookmarkfolder.query.infrastructure.repository.FindFolderRepository;
-import com.runninghi.common.annotation.DomainService;
 import com.runninghi.feedback.command.domain.exception.customException.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
 @RequiredArgsConstructor
 @Service
-public class FindBookmarkFolderService {
+public class QueryBookmarkFolderService {
 
     private final FindFolderRepository findFolderRepository;
